requirements.txt
aiohttp>=3.9.0
websockets>=12.0,<14.0
aiosqlite>=0.19.0
pyyaml>=6.0.1
rich>=13.0.0